Updated Video.java to use String for Id.
MongoDB's auto generated ObjectId consists of alphanumeric characters and hyphens '-'. When a long is used as the datatype for the Id, Spring Data MongoDB will issue an update/overwrite a single document in MongoDB with '_Id': NumberLong(0). This differs from the desired outcome which is to create a new (additional) document on subsequent calls to src / test / java / org / magnum / mobilecloud / integration / test / VideoSvcClientApiTest.java